Move ColorMode enum to mapping package
